requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
pandas>=1.3.0
python-dotenv>=0.19.0
flask>=2.0.0
asyncio>=3.4.3
pytest>=6.2.0
pytest-asyncio>=0.18.0